Fixes #84
To the best of my knowledge, the NOTE generated that caused submission to fail was because I submitted the entire text of the MIT license like every other sane language does. But because CRAN is run by a bunch of assholes, they have to be opaque and different, and thus, I have to guess that this will fix the error based on how Hadley documented his httr package.